app/requirements.txt
bottle==0.13.2
mysql-connector-python==9.2.0
